Extract variance and correlation components

Extract variance and correlation components

# S3 method for glmmTMB
VarCorr(x, sigma = 1, ...)

Arguments

x
a fitted glmmTMB model

sigma
residual standard deviation (usually set automatically from internal information)

extra
arguments (for consistency with generic method)

Details

For an unstructured variance-covariance matrix, the internal parameters are structured as follows: the first n parameters are the log-standard-deviations, while the remaining n(n-1)/2 parameters are the elements of the Cholesky factor of the correlation matrix, filled in column-wise order (see the TMB documentation for further details).

Examples

## Comparing variance-covariance matrix with manual computation
data("sleepstudy",package="lme4")
fm4 <- glmmTMB(Reaction ~ Days + (Days|Subject), sleepstudy)
VarCorr(fm4)[[c("cond","Subject")]]

#>             (Intercept)     Days
#> (Intercept)   565.51573 11.05626
#> Days           11.05626 32.68204
#> attr(,"stddev")
#> (Intercept)        Days 
#>   23.780575    5.716821 
#> attr(,"correlation")
#>             (Intercept)       Days
#> (Intercept)  1.00000000 0.08132635
#> Days         0.08132635 1.00000000
#> attr(,"blockCode")
#> us 
#>  1 

## hand calculation
pars <- getME(fm4,"theta")
## construct cholesky factor
L <- diag(2)
L[lower.tri(L)] <- pars[-(1:2)]
C <- crossprod(L)
diag(C) <- 1
sdvec <- exp(pars[1:2])
(V <- outer(sdvec,sdvec) * C)

#>          [,1]     [,2]
#> [1,] 565.5157 11.09300
#> [2,]  11.0930 32.68204