Extract fixed-effects estimates

Extract Fixed Effects

# S3 method for glmmTMB
fixef(object, ...)

Arguments

object

any fitted model object from which fixed effects estimates can be extracted.

...

optional additional arguments. Currently none are used in any methods.

Value

an object of class fixef.glmmTMB comprising a list of components (cond, zi, disp), each containing a (possibly zero-length) numeric vector of coefficients

Details

Extract fixed effects from a fitted glmmTMB model.

The print method for fixef.glmmTMB object only displays non-trivial components: in particular, the dispersion parameter estimate is not printed for models with a single (intercept) dispersion parameter (see examples)

Examples

data(sleepstudy, package = "lme4")
fm1 <- glmmTMB(Reaction ~ Days, sleepstudy)
(f1 <- fixef(fm1))
#> 
#> Conditional model:
#> (Intercept)         Days  
#>      251.41        10.47  
f1$cond
#> (Intercept)        Days 
#>   251.40511    10.46728 
## show full coefficients, including empty z-i model and
## constant dispersion parameter
print(f1, print_trivials = TRUE)
#> 
#> Conditional model:
#> (Intercept)         Days  
#>      251.41        10.47  
#> 
#> Zero-inflation model:
#> character(0)
#> 
#> Dispersion model:
#> (Intercept)  
#>       7.719