List model options that glmmTMB knows about

List model options that glmmTMB knows about

getCapabilities(what = "all", check = FALSE)

Arguments

what
(character) which type of model structure to report on ("all","family","link","covstruct")

check
(logical) do brute-force checking to test whether families are really implemented (only available for what="family")

Value

if check==FALSE, returns a vector of the names (or a list of name vectors) of allowable entries; if check==TRUE, returns a logical vector of working families

Note

these are all the options that are defined internally; they have not necessarily all been implemented (FIXME!)