Calculate random effect structure Calculates number of random effects, number of parameters, block size and number of blocks. Mostly for internal use.

Calculate random effect structure Calculates number of random effects, number of parameters, block size and number of blocks. Mostly for internal use.

getReStruc(reTrms, ss = NULL, aa = NULL, reXterms = NULL, fr = NULL)

Arguments

reTrms

random-effects terms list

ss

a character string indicating a valid covariance structure. Must be one of names(glmmTMB:::.valid_covstruct); default is to use an unstructured variance-covariance matrix ("us") for all blocks).

aa

additional arguments (i.e. rank)

reXterms

terms objects corresponding to each RE term

fr

model frame

Value

a list

blockNumTheta

number of variance covariance parameters per term

blockSize

size (dimension) of one block

blockReps

number of times the blocks are repeated (levels)

covCode

structure code

Examples

data(sleepstudy, package="lme4")
rt <- lme4::lFormula(Reaction~Days+(1|Subject)+(0+Days|Subject),
                    sleepstudy)$reTrms
rt2 <- lme4::lFormula(Reaction~Days+(Days|Subject),
                    sleepstudy)$reTrms
getReStruc(rt)
#> $`1 | Subject`
#> $`1 | Subject`$blockReps
#> [1] 18
#> 
#> $`1 | Subject`$blockSize
#> [1] 1
#> 
#> $`1 | Subject`$blockNumTheta
#> [1] 1
#> 
#> $`1 | Subject`$blockCode
#> us 
#>  1 
#> 
#> 
#> $`0 + Days | Subject`
#> $`0 + Days | Subject`$blockReps
#> [1] 18
#> 
#> $`0 + Days | Subject`$blockSize
#> [1] 1
#> 
#> $`0 + Days | Subject`$blockNumTheta
#> [1] 1
#> 
#> $`0 + Days | Subject`$blockCode
#> us 
#>  1 
#> 
#>