use of priors in glmmTMB

glmmTMB can accept prior specifications, for doing maximum a posteriori (MAP) estimation (or Hamiltonian MC with the tmbstan package), or outside of a strictly Bayesian framework for the purposes of regularizing estimates

Details

The priors argument to glmmTMB must (if not NULL) be a data frame with columns (at least) prior (character; the prior specification, e.g. "normal(0,2)"); class (the name of the underlying parameter vector on which to impose the prior ("beta", "betazi", "betad", "theta", "thetazi", "psi"); coef a string specifying the particular elements of the parameter vector to apply the prior to (not yet implemented; must be specified as an empty string). At present priors can only be imposed jointly on all of the elements of a specified parameter vector, e.g. all fixed-effect coefficients. The tools in brms for specifying priors (e.g. set_prior should work to produce legal specifications. The available prior distributions are "normal" (mean/sd parameterization); "t" (mean/sd/df); "cauchy" (location/scale); "gamma" (mean/shape). The first three are typically used for fixed effect parameters; the last is typically used for standard deviation parameters ...

to be continued ... (complete separation, singularity, etc.)