Extract residual standard deviation or dispersion parameter

For Gaussian models, sigma returns the value of the residual standard deviation; for other families, it returns the dispersion parameter, however it is defined for that particular family. See details for each family below.

# S3 method for glmmTMB
sigma(object, ...)

Arguments

object

a “glmmTMB” fitted object

...

(ignored; for method compatibility)

Details

The value returned varies by family:

gaussian

returns the maximum likelihood estimate of the standard deviation (i.e., smaller than the results of sigma(lm(...)) by a factor of (n-1)/n)

nbinom1

returns a dispersion parameter (usually denoted \(\alpha\) as in Hardin and Hilbe (2007)): such that the variance equals \(\mu(1+\alpha)\).

nbinom2

returns a dispersion parameter (usually denoted \(\theta\) or \(k\)); in contrast to most other families, larger \(\theta\) corresponds to a lower variance which is \(\mu(1+\mu/\theta)\).

Gamma

Internally, glmmTMB fits Gamma responses by fitting a mean and a shape parameter; sigma is estimated as (1/sqrt(shape)), which will typically be close (but not identical to) that estimated by stats:::sigma.default, which uses sqrt(deviance/df.residual)

beta

returns the value of \(\phi\), where the conditional variance is \(\mu(1-\mu)/(1+\phi)\) (i.e., increasing \(\phi\) decreases the variance.) This parameterization follows Ferrari and Cribari-Neto (2004) (and the betareg package):

betabinomial

This family uses the same parameterization (governing the Beta distribution that underlies the binomial probabilities) as beta.

genpois

returns the index of dispersion \(\phi^2\), where the variance is \(\mu\phi^2\) (Consul & Famoye 1992)

compois

returns the value of \(1/\nu\), When \(\nu=1\), compois is equivalent to the Poisson distribution. There is no closed form equation for the variance, but it is approximately undersidpersed when \(1/\nu <1\) and approximately oversidpersed when \(1/\nu >1\). In this implementation, \(\mu\) is exactly the mean (Huang 2017), which differs from the COMPoissonReg package (Sellers & Lotze 2015).

tweedie

returns the value of \(\phi\), where the variance is \(\phi\mu^p\). The value of \(p\) can be extracted using family_params

The most commonly used GLM families (binomial, poisson) have fixed dispersion parameters which are internally ignored.

References

Consul PC, and Famoye F (1992). "Generalized Poisson regression model. Communications in Statistics: Theory and Methods" 21:89–109.

Ferrari SLP, Cribari-Neto F (2004). "Beta Regression for Modelling Rates and Proportions." J. Appl. Stat. 31(7), 799-815.

Hardin JW & Hilbe JM (2007). "Generalized linear models and extensions." Stata press.

Huang A (2017). "Mean-parametrized Conway–Maxwell–Poisson regression models for dispersed counts. " Statistical Modelling 17(6), 1-22.

Sellers K & Lotze T (2015). "COMPoissonReg: Conway-Maxwell Poisson (COM-Poisson) Regression". R package version 0.3.5. https://CRAN.R-project.org/package=COMPoissonReg