Change starting parameters, either by residual method or by user input (start)

Change starting parameters, either by residual method or by user input (start)

startParams(
  parameters,
  formula,
  ziformula,
  dispformula,
  fr,
  yobs,
  weights,
  size = NULL,
  Xdisp = NULL,
  XdispS = NULL,
  family,
  condReStruc,
  start = NULL,
  sparseX = NULL,
  start_method = list(method = NULL, jitter.sd = 0)
)

Arguments

formula

current formula, containing both fixed & random effects

ziformula

a one-sided (i.e., no response variable) formula for zero-inflation combining fixed and random effects: the default ~0 specifies no zero-inflation. Specifying ~. sets the zero-inflation formula identical to the right-hand side of formula (i.e., the conditional effects formula); terms can also be added or subtracted. When using ~. as the zero-inflation formula in models where the conditional effects formula contains an offset term, the offset term will automatically be dropped. The zero-inflation model uses a logit link.

dispformula

a one-sided formula for dispersion containing only fixed effects: the default ~1 specifies the standard dispersion given any family. The argument is ignored for families that do not have a dispersion parameter. For an explanation of the dispersion parameter for each family, see sigma. The dispersion model uses a log link. In Gaussian mixed models, dispformula=~0 fixes the residual variance to be 0 (actually a small non-zero value), forcing variance into the random effects. The precise value can be controlled via control=glmmTMBControl(zero_dispval=...); the default value is sqrt(.Machine$double.eps).

fr

model frame

yobs

observed y

weights

model weights (for binomial-type models, used as size/number of trials)

size

number of trials in binomial and betabinomial families

family

family object

start

starting values, expressed as a list with possible components beta, betazi, betadisp (fixed-effect parameters for conditional, zero-inflation, dispersion models); b, bzi (conditional modes for conditional and zero-inflation models); theta, thetazi (random-effect parameters, on the standard deviation/Cholesky scale, for conditional and z-i models); psi (extra family parameters, e.g., shape for Tweedie models).

sparseX

see glmmTMB

start_method

Options to initialise the starting values for rr parameters; jitter.sd adds variation to the starting values of latent variables when start = "res".