Repeated counts of salamanders in streams

A data set containing counts of salamanders with site covariates and sampling covariates. Each of 23 sites was sampled 4 times. When using this data set, please cite Price et al. (2016) as well as the Dryad data package (Price et al. 2015).

data(Salamanders)

Format

A data frame with 644 observations on the following 10 variables:

site

name of a location where repeated samples were taken

mined

factor indicating whether the site was affected by mountain top removal coal mining

cover

amount of cover objects in the stream (scaled)

sample

repeated sample

DOP

Days since precipitation (scaled)

Wtemp

water temperature (scaled)

DOY

day of year (scaled)

spp

abbreviated species name, possibly also life stage

count

number of salamanders observed

References

Price SJ, Muncy BL, Bonner SJ, Drayer AN, Barton CD (2016) Effects of mountaintop removal mining and valley filling on the occupancy and abundance of stream salamanders. Journal of Applied Ecology 53 459--468. doi:10.1111/1365-2664.12585

Price SJ, Muncy BL, Bonner SJ, Drayer AN, Barton CD (2015) Data from: Effects of mountaintop removal mining and valley filling on the occupancy and abundance of stream salamanders. Dryad Digital Repository. doi:10.5061/dryad.5m8f6

Examples

require("glmmTMB")
data(Salamanders)
# \donttest{
zipm3 = glmmTMB(count~spp * mined + (1|site), zi=~spp * mined, Salamanders, family="poisson")
# }